move _get_role_edition out of utils
